(PDB-2188) Rename artifact to `puppet-db` and remove paging flags
This commit renames the artifact executable to `puppet-db` to match the
intended artifact name at the time of this commit. In addition this
commit changes the usage of the tool to require the `query` subcommand.
For example:
```bash
puppet-db query '["from","nodes"]
```
This commit also removes the `limit` and `order-by` paging option flags
as they will be able to be specified in the query as part of PQL work.