(maint) Add better config file handling
This commit removes a bug where the default configuration JSON was in
the incorrect format. This commit also adds the ability to pass in
either a single `server_urls` as a string or an array of `server_urls`
for PuppetDB failover. This commit catches JSON parse errors as well and
warns the user when they are encoutered, falling back to the default
configuration.

This commit also aligns the format of the CLI configuration file with
all of the other PuppetDB configuration files, such that there is now a
single `puppetdb` header where you specify your server_urls and ssl
credentials. We do lose the ability to specify multiple PuppetDB
environments but we'll gain this back we add in the ability to specify
the path to the config file via a command line flag.