(maint) Make a PuppetDBConn class for holding configuration
This commit makes a PuppetDBConn class for holding PuppetDB
configuration which makes it easier implementationwise to provide errors
to the user and will provide a means for easily dealing with multiple
server_urls in the config eventually.

This commit also renames our namespace and library from puppetdb_cli and
libpuppetdb_cli to puppetdb and libpuppetdb, so that we might live in a
saner world.