(maint) Fix error messages for bad requests
This commit fixes error messages for requests to PuppetDB that were
executed successfully but which returned a non-200 status code, such as
parse errors in PQL etc. Prior to this commit these requests returned
the PQL error and then returned a `No error found` message because we
we're checking for curl execution errors.