(PDB-2815) Look for global configuration file
This commit adds the ability to use a global configuration file with the
PuppetDB CLI, which we will fall back to if values in the per-user
config file are missing. This global config file is located at
`/etc/puppetlabs/client-tools/puppetdb.conf` similar to the location for
the global config for the other pe-client-tools.