(PE-18967) Add PE_FAMILY env for acceptance tests
This commit adds a `PE_FAMILY` environment variable to be used in
the creation of the `pe_dist_dir` and `pe_ver` values. The existing
default value for `pe_dist_dir` has been removed. This means that the
`PE_FAMILY` value must be set by the Jenkins job or the rake task will fail.