(maint) Fix manpages generation with some versions of Perl
Recent Perl (v5.26.1) on FreeBSD seems to not accept not passing a value
to it's --release argument.  As a consequence, "--center" was considered
to be the release, "$center" was considered to be the pod file to read,
"--name" and "$manfile" where properly handled, and "man/${manfile}.pod"
was considered to be the file to generate and was cleared :-/

Explicitly pass an empty release to unbreak man pages generation.