(maint) Fix token handling again
Provide a more specific error message when a user-supplied token file
does not exist.

Also, cobra provides the Changed function which tells us if a flag was
specified by the user. This is a better check than comparing the flag
with the default value (as the user can explicitly pass in the default
value as well).